requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
PyYAML
pytest
httpx
pytest-mock